requirements.txt
flask
gunicorn
groq
Pillow
python-dotenv